Deleting, Creating, and Editing works for schools

+++ backend/routes/schools.js
@@ -1,5 +1,6 @@
 import express from 'express';
-import { getAllSchools, getSpecificSchool, deleteSchool } from '../services/schoolsService.js';
+import { getAllSchools, getSpecificSchool, deleteSchool, createSchool, editSchool } from '../services/schoolsService.js';
+import { authenticateToken } from '../middleware/auth.js';
 
 const router = express.Router();
 
@@ -35,16 +36,65 @@ router.get('/:school_id', async (req, res) => {
     }
 })
 
-router.delete('/:school_id', async (req, res) => {
+
+/**
+ * ADMIN and Protected Endpoints
+ */
+router.delete('/:school_id', authenticateToken, async (req, res) => {
     try {
         const { school_id } = req.params;
+
+        if(!req.user) {
+            return res.status(401).json({ error: "Authentication Required"})
+        }
     
         await await deleteSchool(req.pgClient, school_id);
-        res.status(201).json({ message: 'School successfully deleted'})
+        res.status(200).json({ message: 'School successfully deleted'})
     } catch (error) {
         console.error('Error Deleting school data', error);
         res.status(500).json({ error: error.message});
     }
 });
+
+router.post('/', authenticateToken, async (req, res) => {
+    try {
+        const schoolData = req.body;
+
+        console.log(schoolData);
+
+        if (!req.user) {
+            return res.status(401).json({ error: 'Authentication required' });
+        }
+        
+        const { username } = req.user;
+        console.log("In backend: ", username);
+
+        const result = await createSchool(req.pgClient, schoolData);
+        res.status(201).json({ message: 'School Successfully Created' });
+    } catch (error) {
+        console.error('Error creating school:', error);
+        res.status(500).json({ error: error.message });
+    }
+})
+
+router.put('/:school_id', authenticateToken, async (req, res) => {
+    try {
+        const schoolData = req.body;
+        console.log(schoolData);
+
+        if (!req.user) {
+            return res.status(401).json({ error: 'Authentication required' });
+        }
+        
+        const { username } = req.user;
+        console.log("In backend: ", username);
+
+        const result = await editSchool(req.pgClient, schoolData);
+        res.status(200).json({ message: 'School Successfully Updated' });
+    } catch (error) {
+        console.error('Error editing school:', error);
+        res.status(500).json({ error: error.message });
+    }
+})
 
 export default router;
